Extract tag name and status enum
