+++ src/StackStatusTag.ts
@@ -1,0 +1,6 @@
+export const TagName = "stackjanitor";
+
+export enum StackStatus {
+  Enabled = "enabled",
+  Disabled = "disabled"
+}
